+++ src/__tests__/hooks/medium.useNotifications.spec.ts
@@ -2,8 +2,6 @@ import { act, renderHook } from '@testing-library/react';
 
 import { useNotifications } from '../../hooks/useNotifications.ts';
 import { Event } from '../../types.ts';
-import { formatDate } from '../../utils/dateUtils.ts';
-import { parseHM } from '../utils.ts';
 
 describe('useNotifications', () => {
   const mockEvents: Event[] = [
