+++ src/const/calendarOptions.ts
@@ -1,0 +1,3 @@
+export const categories = ['업무', '개인', '가족', '기타'];
+
+export const weekDays = ['일', '월', '화', '수', '목', '금', '토'];
